### 2025-10-15 #9258 @editorialbot
[PRE REVIEW]: VowSpace: A vowel formant analysis application for phonetic research
**Submitting author:** @alicagank (<a href="http://orcid.org/0009-0007-9386-9766">Ali Çağan Kaya</a>)
**Repository:** https://github.com/alicagank/VowSpace
**Branch with paper.md** (empty if default branch):
**Version:** v1.4.2
