### 2025-11-05 8424232
Update paper.md and references, added missing DOIs
