+++ paper/paper.bib
@@ -60,7 +60,8 @@
   year    = {2018},
   version = {1.2-2},
   howpublished = {\url{https://cran.r-project.org/web/packages/vowels/vowels.pdf}},
-  note    = {[R Package]}
+  note    = {[R Package]},
+  doi = {10.32614/CRAN.package.vowels}
 }
 
 @article{lobanov1971,
@@ -79,7 +80,8 @@
   year    = {2016},
   version = {1.0-7},
   howpublished = {\url{https://cran.r-project.org/web/packages/phonR/phonR.pdf}},
-  note    = {[R Package]}
+  note    = {[R Package]},
+  doi = {10.32614/CRAN.package.phonR}
 }
 
 @article{moore1983,
@@ -102,20 +104,21 @@
 }
 
 @article{parselmouth2018,
-    author = "Yannick Jadoul and Bill Thompson and Bart de Boer",
-    title = "Introducing {P}arselmouth: A {P}ython interface to {P}raat",
-    journal = "Journal of Phonetics",
-    volume = "71",
-    pages = "1--15",
-    year = "2018",
-    doi = "https://doi.org/10.1016/j.wocn.2018.07.001"
+    author = {Yannick Jadoul and Bill Thompson and Bart de Boer},
+    title = {Introducing {P}arselmouth: A {P}ython interface to {P}raat},
+    journal = {Journal of Phonetics},
+    volume = {71},
+    pages = {1--15},
+    year = {2018},
+    doi = {10.1016/j.wocn.2018.07.001}
 }
 
 @manual{r2021,
   author    = {{R Core Team}},
   title     = {R: A language and environment for statistical computing},
   organization = {R Foundation for Statistical Computing},
-  year      = {2021}
+  year      = {2021},
+  doi       = {10.32614/R.manuals}
 }
 
 @online{remirez2022,
@@ -148,11 +151,12 @@
 }
 
 @misc{thomas2007,
-  author = {Thomas, Erik R. and Kendall, Tyler},
-  title  = {NORM: The vowel normalization and plotting suite},
-  year   = {2007},
-  howpublished = {\url{http://ncslaap.lib.ncsu.edu/tools/norm/}},
-  note   = {[Web application]}
+  author    = {Thomas, Erik R. and Kendall, Tyler},
+  title     = {NORM: The vowel normalization and plotting suite},
+  year      = {2007},
+  howpublished = {Online resource},
+  url       = {http://ncslaap.lib.ncsu.edu/tools/norm/},
+  note      = {[Web application]}
 }
 
 @article{traunmuller1990,
