+++ paper/paper.bib
@@ -1,12 +1,16 @@
 @article{adank2004,
-  author  = {Adank, Patti and Smits, Roel and van Hout, Roeland},
-  title   = {A comparison of vowel normalization procedures for language variation research},
-  journal = {Journal of the Acoustical Society of America},
-  volume  = {116},
-  number  = {5},
-  pages   = {3099--3017},
-  year    = {2004},
-  doi     = {10.1121/1.1795335}
+    author = {Adank, Patti and Smits, Roel and van Hout, Roeland},
+    title = {A comparison of vowel normalization procedures for language variation research},
+    journal = {The Journal of the Acoustical Society of America},
+    volume = {116},
+    number = {5},
+    pages = {3099-3107},
+    year = {2004},
+    month = {11},
+    issn = {0001-4966},
+    doi = {10.1121/1.1795335},
+    url = {https://doi.org/10.1121/1.1795335},
+    eprint = {https://pubs.aip.org/asa/jasa/article-pdf/116/5/3099/8094222/3099_1_online.pdf},
 }
 
 @software{boersma2025,
@@ -130,7 +134,6 @@
   doi = {10.5281/zenodo.7232005},
   langid = {en}
 }
-
 
 @book{stevens1998,
   author    = {Stevens, Kenneth N.},
