### 2026-01-15 #9258 @arfon
### Update regarding JOSS scope and paper format changes

:wave:@alicagank. Thank you for your patience while your submission has been waiting in our pre-review queue.

#### What this means for your submission

- Software Design (~200 words): Explain the trade-offs you weighed, the design/architecture you settled on, and why it matters. If other packages exist in this space, you must include an explicit "build vs. contribute" justification explaining why you created new software rather than contributing to existing projects.
- Research Impact Statement (~200 words): Provide concise evidence of either realized impact (e.g., external use, integrations, enabled results) or credible near-term significance (novel capability with benchmarks, reproducible materials, and community-readiness signals such as documentation, tests, license, releases, and contribution process).
- AI Usage Disclosure: Disclose any use of generative AI in your submission, including software creation, documentation generation, or paper authoring. If you did not use AI tools, please state this explicitly.

For full details, please see:

- Our updated scope definition: https://joss.readthedocs.io/en/latest/submitting.html#scope-and-significance
- An example paper in the new format: https://joss.readthedocs.io/en/latest/example_paper.html

**Important:** Once you have updated your paper, please comment here to let us know. At that point, we will assess your software against our updated scope criteria. We want to be transparent that under these new criteria, some submissions that were previously in scope may now be deemed out of scope, and we may choose not to proceed with review.
