### 2026-01-15 #9258 @alicagank
@arfon Thank you very much for the update, your guidence, and for the detailed explanation! I will update the paper to match the revised format and scope criteria as soon as possible. I will comment here once the changes are complete. :)
