### 2026-01-27 c96202a
Update paper.md in accordance with JOSS's new scope
