+++ paper/paper.md
@@ -17,50 +17,55 @@ date: 4 July 2025
 bibliography: paper.bib
 ---
 
-# Background
+# Summary
 
 Vowel formant frequencies, which vary by vowel and speaker, are crucial in phonetic and sociolinguistic studies. These frequencies reveal language- and speaker-specific traits [@clopper2005]. F1 and F2 relate to vowel height and backness, aiding in vowel distinction, while F3, F4, and F5 provide more detailed speaker information [@fant1960; @stevens1998]. Measuring these frequencies is key across phonetics, sociolinguistics, dialectology, speech technology, and language learning.
 
 Formant values are affected by a speaker’s vocal tract length, gender, and age, making comparisons difficult. To address this, researchers have developed various vowel normalization techniques to reduce physiological differences while preserving phonetic contrasts [@adank2004; @disner1980; @lobanov1971].
 
-Visualization of formant values through F1-F2 scatterplots facilitates the analysis of vowel inventories, regional and diachronic differences, and speech perception experiments. However, these visualizations often require complex scripting or software, limiting access for non-specialists or students.
+Visualization of formant values through F1-F2 scatterplots facilitates the analysis of vowel inventories, regional and diachronic differences, and speech perception experiments. However, these processes often require complex scripting or software, limiting access for non-specialists or students.
 
-# The VowSpace application
+# Statement of need
 
-VowSpace is an open-source desktop application for acquiring, visualizing, normalizing, comparing, and analyzing vowel formant frequencies from audio files and/or datasets. It features an intuitive graphical user interface for exploring vowel space across speakers and groups, with normalization methods including Bark Difference [@traunmuller1990], Lobanov [@lobanov1971; @remirez2022], Nearey 1 and 2 [@nearey1978], Bark [@traunmuller1990], Log, Mel [@stevens1940], and Erb [@moore1983]. It helps researchers visualize vowels, identify patterns, and compare formants. 
+In phonetic and sociophonetic research, analyzing vowel formants is essential for understanding language variation, speaker physiology, and vowel space. Researchers often use general-purpose software, such as R [@r2021] or Praat [@boersma2025], which require advanced skills and time for custom visualizations. VowSpace is an open-source tool for vowel space plotting, combining raw and normalized data. While R packages like phonR [@mccloy2016] and vowels [@kendall2018], as well as Praat, offer similar functions, they either assume programming expertise or lack accessible GUI features for normalization and visualization.
 
-VowSpace enables researchers to load, normalize, and visualize formant data, providing publication-ready results across platforms that support interactive, reproducible workflows.
+Besides R packages like phonR and vowels, and the widely used desktop application Praat, there are two web-based tools for vowel formant data normalization and visualization. NORM [@thomas2007] supports multiple normalization algorithms, including Lobanov, Nearey, Watt & Fabricius, and Bark difference, and allows for vowel space plotting and data export for cross-speaker and dialect comparisons. Visible Vowels [@heeringa2018] offers an interactive platform for uploading formant data, creating vowel plots, and conducting exploratory analyses such as clustering and multidimensional scaling, focusing on visual analytics for phonetic and sociolinguistic research.
 
-![VowSpace's main user interface](jossimage1.png)
[…]
