### 2026-01-27 c7206d3
Update paper.md and paper.bib in accordance with JOSS's new scope
