+++ paper/paper.md
@@ -13,7 +13,7 @@ authors:
 affiliations:
   - name: Hacettepe University, Department of English Linguistics, Ankara, Türkiye
     index: 1
-date: 4 July 2025
+date: 27 January 2026
 bibliography: paper.bib
 ---
 
