+++ paper/paper.bib
@@ -50,13 +50,15 @@
   year      = {1960}
 }
 
-@misc{heeringa2018,
-  author = {Heeringa, Wilbert J. and Van de Velde, Hans},
-  title  = {Visible vowels},
-  year   = {2018},
-  howpublished = {\url{http://www.visiblevowels.org}},
-  note   = {[Web application]}
+@inproceedings{heeringa2018,
+  author    = {Heeringa, Wilbert J. and Van de Velde, Hans},
+  title     = {Visible Vowels: A Tool for the Visualization of Vowel Variation},
+  booktitle = {Proceedings of the CLARIN Annual Conference 2018},
+  year      = {2018},
+  address   = {Pisa, Italy},
+  publisher = {CLARIN ERIC},
 }
+
 
 @misc{kendall2018,
   author  = {Kendall, Tyler and Thomas, Erik R.},
