### 2026-01-28 b5e9170
Update paper.md in accordance with JOSS's new scope
