+++ paper/paper.md
@@ -29,6 +29,10 @@ Visualization of formant values through F1-F2 scatterplots facilitates the analy
 
 In phonetic and sociophonetic research, analyzing vowel formants is essential for understanding language variation, speaker physiology, and vowel space. Researchers often use general-purpose software, such as R [@r2021] or Praat [@boersma2025], which require advanced skills and time for custom visualizations. VowSpace is an open-source tool for vowel space plotting, combining raw and normalized data. While R packages like phonR [@mccloy2016] and vowels [@kendall2018], as well as Praat, offer similar functions, they either assume programming expertise or lack accessible GUI features for normalization and visualization.
 
+VowSpace was developed to fill this gap by providing an offline desktop application for acquiring, handling, normalizing, and visualizing vowel formant data within a single environment. It was designed to minimize the level of technical expertise required to work with vowel formants, thereby making this area of research more accessible to newcomers. Considering the lack of a dedicated application for vowel formant analysis and the requirements of fields such as forensic phonetics, which necessitate offline platforms, VowSpace can serve as a primary tool for a wide range of phonetic analyses.
+
+# State of the field
+
 Besides R packages like phonR and vowels, and the widely used desktop application Praat, there are two web-based tools for vowel formant data normalization and visualization. NORM [@thomas2007] supports multiple normalization algorithms, including Lobanov, Nearey, Watt & Fabricius, and Bark difference, and allows for vowel space plotting and data export for cross-speaker and dialect comparisons. Visible Vowels [@heeringa2018] offers an interactive platform for uploading formant data, creating vowel plots, and conducting exploratory analyses such as clustering and multidimensional scaling, focusing on visual analytics for phonetic and sociolinguistic research.
 
 VowSpace offers advantages over web-based tools like NORM and Visible Vowels, making it suitable for large-scale or advanced phonetic research. As a desktop application, it avoids browser and internet limitations, offline data security, and offers comprehensive normalization and visualization options. It integrates normalization methods, along with interactive plotting features like zoomable vowel charts, customizable colors, and high-quality exports. Additionally, VowSpace provides simple data reading and writing, graphical plotting with speaker metadata, exportable figures suitable for publication, and audio analysis tools for vowel formant acquisition through spectrograms and formant lines. These capabilities make VowSpace a flexible, precise, and user-friendly alternative for research and teaching across phonetics, phonology, sociolinguistics, and language teaching and acquisition.
