### 2026-01-30 adc7ca5
I forgot to add openpyxl as a requirement for Excel i/o
