+++ requirements.txt
@@ -4,3 +4,4 @@ PyQt5~=5.15.10
 matplotlib~=3.8.3
 scipy~=1.12.0
 praat-parselmouth>=0.4
+openpyxl
