### 2026-01-30 e84e5c6
I forgot to add openpyxl as a requirement for Excel i/o
