### 2026-03-08 #9258 @RichardLitt
@zc-guo, @hadware, and @jjatria: Would any of you be interested in reviewing this paper? You were the only three [reviewers](https://reviewers.joss.theoj.org/reviewers/search?language=&area_id=&keywords=phonetics&name=&commit=Search) with the word 'phonetics' in your research scope.
