### 2026-03-15 #9258 @RichardLitt
That's OK! Thank you. Let me close out this Pre-review thread and open up another one, where you can make your checklist. This will help: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
