### 2026-03-15 #10189 @editorialbot
[REVIEW]: VowSpace: A vowel formant analysis application for phonetic research
**Submitting author:** @alicagank (<a href="http://orcid.org/0009-0007-9386-9766">Ali Çağan Kaya</a>)
**Repository:** https://github.com/alicagank/VowSpace
**Branch with paper.md** (empty if default branch):
**Version:** 1.4.4
