### 2026-03-15 #9258 @elenlefoll
> [@FredericBlum](https://github.com/FredericBlum) [@elenlefoll](https://github.com/elenlefoll) Would you be interested in reviewing this?

I am afraid I am not a phonetics person and am therefore wholly unqualified for this review. Perhaps @samkirkham is up for it?
