### 2026-03-15 #9258 @RichardLitt
@elenlefoll I figured that might be the case - you had "Linguistics" as your area on the JOSS Reviewers board, but it was a stretch.

@samkirkham a third reviewer if possible. I've also asked another person off of GitHub for help if you can't.
