### 2026-03-19 alicagank/vowelspacevisualizer#1 @FredericBlum
Review: Improved community guidelines
Part of the [review](https://github.com/openjournals/joss-reviews/issues/10189) for the paper.

The community guidelines could be improved in order to move the workflow to GitHub as well, instead of moving the forms to a closed application (like google forms). Here are two ideas:

- [x] GitHub Issue/PR Templates
- [x] Using a CONTRIBUTING.md file (https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors)

This would also makte it easier to give people credit if they decide to contribute and to keep everything open in case you decide to abandon the project at some point.
