### 2026-03-19 alicagank/vowelspacevisualizer#2 @FredericBlum
Review: Code tests
Part of the [review](https://github.com/openjournals/joss-reviews/issues/10189) for the paper.

Right now, no code checks are implemented. Please provide a `pyests` (or similar) functionality that checks that all parts of the code are used and work as intended.
