### 2026-03-19 alicagank/vowelspacevisualizer#3 @FredericBlum
Review: Ethics statement for tutorial data
The tutorial data provided needs some information on the sound-files provided. This is in line with the "Human participants research policy", to which the recorded data definitely belongs.

> In the exceptional case a JOSS submission contains original data on human participants, the study must have been performed in accordance with the [Declaration of Helsinki](https://www.wma.net/policies-post/wma-declaration-of-helsinki-ethical-principles-for-medical-research-involving-human-subjects/). The manuscript must contain all relevant information regarding ethics approval, including but not limited to the responsible ethics committee and reference number. This also applies to ethics exemptions. Informed consent must be collected from all human research participants, and authors are required to state whether this occurred in the manuscript.
