### 2026-03-19 alicagank/vowelspacevisualizer#4 @FredericBlum
Review: Possible file formats for import
Neither in the repository nor in the paper did I find an overview of which file types can be imported. Please make this explicit in both.
