### 2026-03-19 alicagank/vowelspacevisualizer#5 @FredericBlum
Review: Software design, depth of description
Overall, the paper is well written. However, I sometimes miss a little bit of depth in the individual sections. Especially the Software Design could need some more content (e.g. #4 ), but also the Summary and the Statement of Need are quite short.

Some ideas:

- explain the visualization methods provided - e.g. I didn't knew the term "Qhull" before
- give more detail on why normalization is necessary
- add input file formats
- give more context on possible use cases - what should I use this for?
