### 2026-03-19 alicagank/vowelspacevisualizer#6 @FredericBlum
Review: Research impact statement
Please cite the studies from the research projects explicitly instead of only naming them (e.g. line 83).
