### 2026-03-19 alicagank/VowSpace#7 @FredericBlum
Review: References
Some minor problems:

- 106: American English
- 107: Issue missng (3)
- 110: Address (this might be a style thing, but normally I'd expect it to be there)
- 118: Russian
- 129: Inconsistent caps
- 131: Address (this might be a style thing, but normally I'd expect it to be there)
- 134: link does not work
