### 2026-03-19 #10189 @FredericBlum
Thanks @RichardLitt for the guidance and to the author for developing this software. Overall, I am very happy with the software and the information provided. However, there are also some points for which I'd propose some further changes to the repository and to the paper. I provide the overview here and create individual issues in the repository for all the topics.

Related to the repository:

- [x] Improved community guidelines (https://github.com/alicagank/VowSpace/issues/1)
- [x] Add code tests and check their coverage (https://github.com/alicagank/VowSpace/issues/2)
- [x] Elaborate on ethics statement for tutorial-data (https://github.com/alicagank/VowSpace/issues/3)
- [x] Name possible file formats for input to software (https://github.com/alicagank/VowSpace/issues/4)

Related to the paper:

- [x] Software design and overall depth of paper (https://github.com/alicagank/VowSpace/issues/5)
- [x] Research Impact Statement: Cite studies explicitly (https://github.com/alicagank/VowSpace/issues/6)
- [x] References: Some small fixes (https://github.com/alicagank/VowSpace/issues/7)
