+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,0 +1,25 @@
+---
+name: Bug report
+about: Report a bug in VowSpace
+title: "[BUG] "
+labels: bug
+---
+
+## Description
+Describe the bug clearly.
+
+## Steps to Reproduce
+1. Data used...
+2. Click on/try...
+3. See error...
+
+## Expected Behavior
+What should happen?
+
+## Screenshots and Crash Reports
+If applicable.
+
+## Environment
+- OS:
+- Python version:
+- VowSpace version:
