+++ .github/ISSUE_TEMPLATE/feature_request.md
@@ -1,0 +1,18 @@
+---
+name: Feature request
+about: Suggest a new feature
+title: "[FEATURE] "
+labels: enhancement
+---
+
+## Problem
+What problem does this solve?
+
+## Proposed Solution
+Describe your idea.
+
+## Alternatives
+Any alternative solutions?
+
+## Additional context
+Anything else?
