+++ CONTRIBUTING.md
@@ -1,0 +1,63 @@
+# Contributing to VowSpace
+
+Thank you for your interest in contributing to VowSpace! :))
+I welcome contributions from researchers, developers, and users alike.
+
+## Ways to Contribute
+
+You can contribute in several ways:
+
+- Reporting bugs
+- Suggesting new features
+- Improving documentation
+- Adding tests and data
+- Contributing code
+
+## Reporting Issues
+
+Please use the GitHub Issues tab to report bugs or request features.
+
+When reporting an issue, please include:
+- A clear description of the problem
+- Steps to reproduce (if applicable)
+- Expected vs actual behavior
+- Screenshots (if relevant)
+- Your system information (OS, Python version)
+
+## Suggesting Features
+
+Feature requests are also welcome!
+Please describe:
+- The problem your feature solves
+- Why it would be useful
+- Possible implementation ideas (optional)
+
+## Contributing Code
+
+1. Fork the repository
+2. Create a new branch (`feature/your-feature-name`)
+3. Make your changes
+4. Make sure code is clean and documented
+5. Add tests if applicable
+6. Submit a Pull Request
+
+## Coding Guidelines
+
+- Follow existing code structure and naming conventions
+- Keep functions modular and readable
+- Document key functions
+- Avoid unnecessary dependencies
+
+## Tests
+
+If you add new functionality, please include tests if possible.
+
+## Code of Conduct
+
+Please be respectful and constructive in all interactions.
+
+## Credit
+
+All contributors will be acknowledged in the project, both on GitHub and on the relevant page on my personal website. :)
+
+Thank you for helping improve VowSpace!
