+++ .github/pull_request_template.md
@@ -1,0 +1,15 @@
+## Description
+What does this PR do?
+
+## Type of Change
+- [ ] Bug fix
+- [ ] New feature
+- [ ] Documentation update
+
+## Checklist
+- [ ] Code follows project style
+- [ ] Tests added/updated
+- [ ] Documentation updated
+
+## Related Issues
+Link issues here if applicable
