### 2026-03-19 0cf3339
Added FLAC and AIFF support for audio reading
