### 2026-03-19 b8fcc29
Documented supported audio and tabular formats
