### 2026-03-19 bd1ec3c
test: added pytest-based test suite with coverage for normalization, IO, GUI, data handling and audio tools
