### 2026-03-19 alicagank/vowelspacevisualizer#1 @alicagank
Thank you for this helpful suggestion.

A CONTRIBUTING.md file has been added to the repository, and the contribution workflow has been moved to GitHub using Issues and Pull Requests (see .github/ISSUE_TEMPLATE/). Templates have also been introduced to standardize contributions and improve transparency.

Implemented in commit 7e265fc.

Closing this issue.
