### 2026-03-19 alicagank/vowelspacevisualizer#4 @alicagank
Thank you for this helpful suggestion.

Support for audio file formats (.aiff, .aif, .aifc, .flac, .wav, .mp3, .ogg) and tabular formats (CSV and Excel) has been implemented in the codebase (see audio_tool.py and vowel_space_visualizer.py) and documented in the README.

These changes were implemented in commits 088374f and b8fcc29.

I will update the manuscript accordingly to explicitly include this information as well, after which I will close this issue.
