### 2026-03-19 alicagank/vowelspacevisualizer#2 @alicagank
Thank you for this important suggestion.

A pytest-based testing framework has now been added to the repository (see tests/). The current tests cover the core normalization and transformation functions, table import and export workflows, the main functionalities of the Audio Analysis Tools window, and the DataFrame Editor. Coverage reporting has also been added using pytest-cov, with the normalization module currently reaching 97% coverage.

While the graphical interface is not exhaustively tested at the level of every individual UI interaction, the principal computational and data-processing components are now verified.

If you have further suggestions for additional tests, I would be very happy to incorporate them.

Implemented in commit bd1ec3c.

Closing this issue.
