+++ paper/paper.bib
@@ -26,7 +26,7 @@
   author  = {Clopper, Cynthia G. and Pisoni, David B. and De Jong, Kenneth},
   title   = {Acoustic characteristics of the vowel systems of six regional varieties of American English},
   journal = {Journal of the Acoustical Society of America},
-  volume  = {118},
+  volume  = {118(3 Pt 1)},
   pages   = {1661--1676},
   year    = {2005},
   doi     = {10.1121/1.2000774}
@@ -47,7 +47,8 @@
   author    = {Fant, Gunnar},
   title     = {Acoustic theory of speech production},
   publisher = {Mouton},
-  year      = {1960}
+  year      = {1960},
+  address   = {The Hague, The Netherlands},
 }
 
 @inproceedings{heeringa2018,
@@ -129,7 +130,7 @@
 
 @online{remirez2022,
   author = {Remirez, Emily},
-  title = {Vowel Plotting in {Python}},
+  title = {Vowel plotting in {Python}},
   series = {Linguistics Methods Hub},
   date = {2022-10-20},
   url = {https://lingmethodshub.github.io/content/python/vowel-plotting-py},
@@ -137,11 +138,15 @@
   langid = {en}
 }
 
-@book{stevens1998,
-  author    = {Stevens, Kenneth N.},
-  title     = {Acoustic phonetics},
-  publisher = {MIT Press},
-  year      = {1998}
+@book{stevens1999,
+    author = {Stevens, Kenneth N.},
+    title = {Acoustic Phonetics},
+    publisher = {The MIT Press},
+    year = {1999},
+    month = {01},
+    isbn = {9780262284387},
+    doi = {10.7551/mitpress/1072.001.0001},
+    url = {https://doi.org/10.7551/mitpress/1072.001.0001},
 }
 
 @article{stevens1940,
