### 2026-03-20 alicagank/vowelspacevisualizer#3 @alicagank
Thank you for this important comment.

An ethics statement has been added to the tutorial documentation (see tutorial/instructions.md) including ethics approval details (Hacettepe University Social Sciences and Humanities Ethics Board, 10 September 2024, Ref: 2024/16) and confirmation of informed consent in accordance with the Declaration of Helsinki in commit 29a5fb2.

In addition, the manuscript has been updated to include an ethics statement in the acknowledgements section in commit 41395db line 75.

Closing this issue.
