### 2026-03-20 alicagank/vowelspacevisualizer#4 @alicagank
I have updated the manuscript to include the supported audio and tabular data formats in commit 41395db.

Closing this issue.
