### 2026-03-20 alicagank/vowelspacevisualizer#5 @alicagank
Thank you for your input and constructive comments.

The manuscript has been revised to provide additional depth across the relevant sections. In particular:

- The visualization methods have been expanded, including clarification of convex hulls (Qhull), ellipses, and grouping options (by speaker or vowel).
- Additional explanation has been added on the role and necessity of normalization in vowel formant research.
- Supported input file formats have been explicitly specified.
- The Statement of Need and Summary sections have been expanded to better contextualize the use cases of the software.

in commit 41395db.

Closing this issue.
