### 2026-03-20 alicagank/vowelspacevisualizer#6 @alicagank
Thank you for your valuable suggestion.

I would like to clarify that VowSpace was developed as the primary technical and analytical tool within the scope of the TÜBİTAK 2209-A project (No: 1919B012414090). As the research associated with these projects is currently in progress or in the final stages of peer-review, there are no prior peer-reviewed publications to cite at this time. The Research Impact Statement has been updated to explicitly state that the software was developed in the course of the project to avoid any confusion regarding its status.

In commit 41395db.

Closing this issue.
