### 2026-03-20 7130e6b
paper: fixed references
