+++ paper/paper.bib
@@ -24,7 +24,7 @@
 
 @article{clopper2005,
   author  = {Clopper, Cynthia G. and Pisoni, David B. and De Jong, Kenneth},
-  title   = {Acoustic characteristics of the vowel systems of six regional varieties of American English},
+  title   = {Acoustic characteristics of the vowel systems of six regional varieties of {A}merican {E}nglish},
   journal = {Journal of the Acoustical Society of America},
   volume  = {118(3 Pt 1)},
   pages   = {1661--1676},
@@ -73,7 +73,7 @@
 
 @article{lobanov1971,
   author  = {Lobanov, Boris M.},
-  title   = {Classification of Russian vowels spoken by different speakers},
+  title   = {Classification of {R}ussian vowels spoken by different speakers},
   journal = {Journal of the Acoustical Society of America},
   volume  = {49},
   pages   = {606--608},
