### 2026-03-20 alicagank/vowelspacevisualizer#7 @alicagank
Thank you for these helpful comments.

The reference list has been revised accordind to your recommendations:

- The citation for Stevens (1998) has been updated to the 1999 MIT Press edition with DOI.
- Capitalization has been corrected (e.g., Lobanov, 1971; Clopper et al., 2005).
- Missing issue information has been added to Clopper et al. (2005).

Regarding Fant (1960), the publisher address has been included in the BibTeX entry, but it is not rendered in the final document possibly due to the CSL citation style, which does not display publisher locations.

Regarding the missing DOIs:

- For Boersma & Weenink (2025), the DOI (10.1097/aud.0b013e31821473f7) refers to a review article published in Ear and Hearing, rather than the software itself. In accordance with the official Praat citation guidelines (https://fon.hum.uva.nl/praat/manual/FAQ__How_to_cite_Praat.html) the software is cited directly, therefore this DOI has not been included.

- For Stevens (1999), the DOI (10.1093/obo/9780199772810-0047) seems to correspond to an Oxford Bibliographies entry written by Allard Jongman rather than Stevens’ book. Therefore the original reference has been retained without adding this DOI.

Regarding the broken link: the citation for Kendall & Thomas (2010) follows the recommended format provided on the official website of the author (see https://lingtools.uoregon.edu/norm/about_norm1.php). The original reference includes the software resource link (http://ncslaap.lib.ncsu.edu/tools/norm/), which appears to no longer be active. For transparency, I have retained the citation format as suggested by the developers. However, I would be happy to remove or replace the link (e.g., with an archived version) if you would prefer.

To ensure that the changes were correctly reflected in the paper PDF, a couple of commits were made. The final version including the relevant changes is contained in commit 7130e6b.

Closing this issue.
