### 2026-03-20 128df5f
paper: fix typos (first usage of GUI) in lines 26–30
