+++ paper/paper.md
@@ -23,11 +23,11 @@ Vowel formant frequencies, which vary by vowel and speaker, are crucial paramete
 
 However, formant values are inherently affected by a speaker’s vocal tract length, gender and age, making comparisons difficult. To address this, researchers have developed various vowel normalization techniques to reduce physiological differences while preserving phonetic contrasts [@adank2004; @disner1980; @lobanov1971]. Such normalization is essential for cross-speaker and cross-dialectal analysis, as it allows for the isolation of linguistic variation from biological variance.
 
-VowSpace facilitates these analyses by providing an integrated environment that supports diverse input formats, including raw audio files (e.g., .wav, .flac, .mp3) for extraction and tabular datasets (e.g., CSV, Excel) for visualization. Visualization of formant values through F1-F2 scatterplots enables the study of vowel inventories, regional and diachronic differences, and speech perception experiments. To provide analytical depth, the application offers multiple visualization methods, including convex hulls and ellipses to delineate category boundaries and overlaps, as well as grouping by vowel or speaker to examine vowel inventories and enable cross-speaker comparisons on a vowel space. The application also offers a comprehensive normalization and conversion suite which includes the Bark Difference Metric [@traunmuller1990], Lobanov [@lobanov1971; @remirez2022], Nearey 1 and 2 [@nearey1978] normalizations, and Bark [@traunmuller1990], Log, Mel [@stevens1940], and Erb [@moore1983] conversions applicable by buttons, implemented as standardized functions to ensure reproducibility across different research settings and longitudinal datasets. While these processes often require complex scripting or specialized software, VowSpace provides an accessible GUI suitable for a wide range of use cases, from classroom instruction to phonetic fieldwork to specialized research in forensic phonetics and language acquisition.
+VowSpace facilitates these analyses by providing an integrated environment that supports diverse input formats, including raw audio files (e.g., .wav, .flac, .mp3) for extraction and tabular datasets (e.g., CSV, Excel) for visualization. Visualization of formant values through F1-F2 scatterplots enables the study of vowel inventories, regional and diachronic differences, and speech perception experiments. To provide analytical depth, the application offers multiple visualization methods, including convex hulls and ellipses to delineate category boundaries and overlaps, as well as grouping by vowel or speaker to examine vowel inventories and enable cross-speaker comparisons on a vowel space. The application also offers a comprehensive normalization and conversion suite which includes the Bark Difference Metric [@traunmuller1990], Lobanov [@lobanov1971; @remirez2022], Nearey 1 and 2 [@nearey1978] normalizations, and Bark [@traunmuller1990], Log, Mel [@stevens1940], and Erb [@moore1983] conversions applicable by buttons, implemented as standardized functions to ensure reproducibility across different research settings and longitudinal datasets. While these processes often require complex scripting or specialized software, VowSpace provides an accessible graphical user interface (GUI) suitable for a wide range of use cases, from classroom instruction to phonetic fieldwork to specialized research in forensic phonetics and language acquisition.
 
 # Statement of need
 
-In phonetic and sociophonetic research, analyzing vowel formants is essential for understanding language variation, speaker physiology, and vowel space. Researchers often use general-purpose software, such as R [@r2021] or Praat [@boersma2025], which require advanced skills and time for custom normalizations and visualizations. This technical barrier can increase the workload because learning complex scripting tools may distract researchers, particularly students and non-specialists, from focusing on the main phonetic analysis. VowSpace is an open-source tool for vowel space plotting, combining raw and normalized data. While R packages like phonR [@mccloy2016] and vowels [@kendall2018], as well as Praat, offer similar functions, they either assume programming expertise or lack accessible graphical user interface (GUI) features for normalization and visualization.
[…]
