### 2026-03-23 ca4f87a
Improved community guidelines, fixed typo
