### 2026-04-17 6dc4735
Refactor project structure and update dependencies. Creating a pyproject.toml file.
