### 2026-04-20 f1bb437
Fixing tests. Added installation instructions based on uv and pip.
