### 2026-04-20 #10189 @hadware
Sorry @RichardLitt : i've picked up review. I'm submitting some PR's to polish the author's already excellent work. Everything should be pretty quick.
