### 2026-04-20 cd5d08c
Add GitHub Actions workflow for testing
