### 2026-04-20 2509bc8
A sprinkle of type hinting
