+++ vowspace/components/df_editor.py
@@ -1,5 +1,3 @@
-# components/df_editor.py
-
 from PyQt5.QtWidgets import QDialog, QTableWidget, QTableWidgetItem, QVBoxLayout, QPushButton
 
 class DFEditor(QDialog):
