### 2026-04-20 a1231c3
Pyproject.toml & Project structure cleanup (#8)
* Refactor project structure and update dependencies. Creating a pyproject.toml file.

* Fixing tests. Added installation instructions based on uv and pip.
