+++ requirements.txt
@@ -1,7 +1,0 @@
-numpy~=1.26.4
-pandas~=2.2.1
-PyQt5~=5.15.10
-matplotlib~=3.8.3
-scipy~=1.12.0
-praat-parselmouth>=0.4
-openpyxl
