### 2026-04-20 2f9aaf4
Automated CI Tests (#10)
* Refactor project structure and update dependencies. Creating a pyproject.toml file.

* Fixing tests. Added installation instructions based on uv and pip.

* Add GitHub Actions workflow for testing
