### 2026-04-20 9f1584e
Trying to fix the test environment setup
