+++ .github/workflows/tests.yml
@@ -21,8 +21,26 @@ jobs:
         with:
           python-version: ${{ matrix.python-version }}
 
+      - name: Install system dependencies
+        run: |
+          sudo apt-get update
+          sudo apt-get install -y \
+            libxcb-xinerama0 \
+            libxcb-icccm4 \
+            libxcb-image0 \
+            libxcb-keysyms1 \
+            libxcb-randr0 \
+            libxcb-render-util0 \
+            libxcb-xkb1 \
+            libxkbcommon-x11-0 \
+            libegl1 \
+            libgl1 \
+            libpulse0
+
       - name: Install dependencies
         run: uv sync --extra dev
 
       - name: Run tests
-        run: uv run xvfb-run pytest
+        env:
+          QT_QPA_PLATFORM: offscreen
+        run: uv run pytest
