### 2026-04-20 #10189 @alicagank
Thank you very much, @hadware, for your help, guidance, and kind words, I really appreciate it!

I’ve reviewed the changes and merged the first PR without any issues; as mentioned there, I had been planning on doing a similar cleanup myself (commit a1231c3).

I also tested the `uv` workflow with Python 3.12, and it runs smoothly on my Fedora system. I plan to test it on other platforms as well, but so far everything looks good. I did encounter an issue with Python 3.14 due to the current `numpy` version, which I plan to address in a future update and release with some additional visualisation methods I am working on.

I did encounter some errors after merging the second PR (commit 2f9aaf4). I attempted a small adjustment to the run command (commit 6306a9d), but I understand you are currently working on a fix for the automated tests, so I’ll wait for your update before proceeding on that.

I’ll hold off on merging the type hinting PR until that is resolved.

Regarding issue #9 (CI for Debian, macOS, and Windows), I have written a workflow for it, but I plan to address it after the current PRs are merged and the project structure is finalised if that is acceptable.

Thanks again for your constructive comments and suggestions!
