### 2026-04-20 bef3cc6
A sprinkle of Type Hinting (#11)
* Refactor project structure and update dependencies. Creating a pyproject.toml file.

* Fixing tests. Added installation instructions based on uv and pip.

* Add GitHub Actions workflow for testing

* A sprinkle of type hinting

* Trying to fix the test environment setup
