+++ .github/workflows/tests.yml
@@ -25,4 +25,4 @@ jobs:
         run: uv sync --extra dev
 
       - name: Run tests
-        run: uv run xvfb-run pytest
+        run: xvfb-run -a uv run pytest
