### 2026-04-21 6fbef2a
Add build workflow
