### 2026-04-21 6fd108d
Add standalone parameter, PtQt5, and icon to the build workflow
