### 2026-04-21 6705e7e
Add standalone parameter, PtQt5, and icon to the build workflow. Fixed parameters.
