+++ .github/workflows/build.yml
@@ -8,13 +8,14 @@ on:
 
 jobs:
   build:
+    name: Build on ${{ matrix.os }}
+    runs-on: ${{ matrix.os }}
+
     strategy:
       fail-fast: false
       matrix:
-        os: [ubuntu-latest]
+        os: [ubuntu-latest, macos-latest, windows-latest]
         python-version: ["3.12"]
-
-    runs-on: ${{ matrix.os }}
 
     steps:
       - name: Check out repository
@@ -24,9 +25,6 @@ jobs:
         uses: actions/setup-python@v5
         with:
           python-version: ${{ matrix.python-version }}
-
-      - name: Install system dependencies (Linux)
-        run: sudo apt-get update && sudo apt-get install -y patchelf
 
       - name: Install project dependencies
         run: |
@@ -38,15 +36,18 @@ jobs:
         with:
           script-name: main.py
           mode: standalone
-          extra-options: >-
-            --enable-plugin=pyqt5
-            --follow-imports
-            --include-data-dir=vowspace/assets=vowspace/assets
+          enable-plugins: pyqt5
+          include-package: PyQt5
+          include-package-data: matplotlib
+          include-data-dir: vowspace/assets=vowspace/assets
+          windows-icon-from-ico: vowspace/assets/vowspace.ico
+          macos-app-icon: vowspace/assets/vowspace.png
+          linux-icon: vowspace/assets/vowspace.png
+          output-dir: build
 
-      - name: Upload build artifacts
+      - name: Upload artifacts
         uses: actions/upload-artifact@v4
         with:
           name: vowspace-${{ runner.os }}-py${{ matrix.python-version }}
           path: |
-            *.dist/**
-            *.bin
+            build/**
