### 2026-04-21 8209937
MacOS standalone -> app
