+++ README.md
@@ -69,7 +69,7 @@ YOU MAY HAVE A LOOK AT THE TUTORIALS I PROVIDED AT [TUTORIAL INSTRUCTIONS](tutor
 
 # Installation
 
-VowSpace is currently available as a portable application.
+VowSpace is currently available as a portable application. The distributions are handled by Nuitka with the build.yml workflow.
 
 ## macOS
 
@@ -86,7 +86,7 @@ VowSpace is currently available as a portable application.
 1. Download the latest `.exe` file from the [GitHub Releases](https://github.com/alicagank/VowSpace/releases) page.
 2. Once downloaded, you can launch **VowSpace**.
    
-## Linux (Debian tested)
+## Linux
 
 1. Download the latest executable file from the [GitHub Releases](https://github.com/alicagank/VowSpace/releases) page.
 2. Once downloaded, you can launch **VowSpace**.
