### 2026-04-21 #10189 @alicagank
Hello @hadware @RichardLitt! I’ve addressed the remaining issue by adding a CI workflow that automatically builds executables for Linux, macOS, and Windows using Nuitka (@hadware's recommendation, which worked out great :)). I have tested the executables on these platforms, and everything seems to be working fine. Implemented in commit 8209937 (added to README in commit d128814). The issue has now been resolved and closed.

Thank you very much for your suggestions and help! I’m happy to address any further issues if there is any; otherwise, I believe everything is complete. ⭐
