### 2026-04-21 69eb02e
Added MacOS .icns file as an icon - this failed Nuitka before. Edited the build.yml to reflect this change.
