### 2026-04-21 94ca02f
Update README.md
Added badges for tests and the build
