### 2026-04-21 93fe2b0
Update paper.md polishing
Changed a claim that sounds a little too strong, typo fixes, consistency of punctuation
