+++ paper/paper.md
@@ -29,7 +29,7 @@ VowSpace facilitates these analyses by providing an integrated environment that 
 
 In phonetic and sociophonetic research, analyzing vowel formants is essential for understanding language variation, speaker physiology, and vowel space. Researchers often use general-purpose software, such as R [@r2021] or Praat [@boersma2025], which require advanced skills and time for custom normalizations and visualizations. This technical barrier can increase the workload because learning complex scripting tools may distract researchers, particularly students and non-specialists, from focusing on the main phonetic analysis. VowSpace is an open-source tool for vowel space plotting, combining raw and normalized data. While R packages like phonR [@mccloy2016] and vowels [@kendall2018], as well as Praat, offer similar functions, they either assume programming expertise or lack accessible GUI features for normalization and visualization.
 
-VowSpace was developed to fill this gap by providing an offline desktop application for acquiring, handling, normalizing, and visualizing vowel formant data within a single environment. It was designed to minimize the level of technical expertise required to work with vowel formants, thereby making this area of research more accessible to newcomers. Considering the lack of a dedicated application for vowel formant analysis and the requirements of fields such as forensic phonetics, which necessitate offline platforms, VowSpace can serve as a primary tool for a wide range of phonetic analyses.
+VowSpace was developed to fill this gap by providing an offline desktop application for acquiring, handling, normalizing, and visualizing vowel formant data within a single environment. It was designed to minimize the level of technical expertise required to work with vowel formants, thereby making this area of research more accessible to newcomers. Considering the limited availability of dedicated applications for vowel formant analysis and the requirements of fields such as forensic phonetics, which necessitate offline platforms, VowSpace can serve as a primary tool for a wide range of phonetic analyses.
 
 # State of the field
 
@@ -41,7 +41,7 @@ VowSpace offers advantages over web-based tools like NORM and Visible Vowels, ma
 
 Developed in Python, VowSpace’s design heavily focuses on ease of use for students and professionals focusing on vowel formants in acquiring, visualizing, normalizing, comparing, and analyzing vowel formant frequencies from audio files and/or datasets. To this end, it features an intuitive GUI for exploring vowel space across speakers and groups with integrated normalization and visualization methods. By providing these tools in a unified environment, VowSpace enables researchers to load, normalize, and visualize formant data, providing publication-ready results across platforms that support interactive, reproducible workflows.
 
-VowSpace has a modular architecture consisting of a main interface and two specialized tools: Audio Analysis and DataFrame Editor. The Audio Analysis window serves as a primary data extraction point, allowing users to load audio files in supported formats (i.e., .aiff, .aif, .aifc, .flac, .wav, .mp3, .ogg), visualize them as spectrograms, and extract acoustic features such as intensity, pitch, and vowel formants (F1–F4) using the Parselmouth library [@parselmouth2018]. Users can also add extracted formant values directly to the visualizer by clicking on the generated spectrogram, which minimizes manual transcription errors and facilitates immediate validation against existing distributions, thereby streamlining the transition from raw signal to analyzed data. These spectrogram images can be saved as JPEG, PNG or TIFF files with visualized acoustic/prosodic information.
[…]
