### 2026-04-21 #10189 @alicagank
Hello @RichardLitt, I’ve given the paper one final read-through to check for typos and made a couple of minor punctuation corrections, and a small wording adjustment where a sentence felt slightly too strong.

I have checked the references as well, I have stated the reasons for the missing DOI numbers at [Issue #7](https://github.com/alicagank/VowSpace/issues/7) opened by FredericBlum.

If there is anything further I should do in this stage, please do let me know. Thank you very much for your help and guidance!
