+++ paper/paper.md
@@ -41,7 +41,7 @@ VowSpace offers advantages over web-based tools like NORM and Visible Vowels, ma
 
 Developed in Python, VowSpace’s design heavily focuses on ease of use for students and professionals focusing on vowel formants in acquiring, visualizing, normalizing, comparing, and analyzing vowel formant frequencies from audio files and/or datasets. To this end, it features an intuitive GUI for exploring vowel space across speakers and groups with integrated normalization and visualization methods. By providing these tools in a unified environment, VowSpace enables researchers to load, normalize, and visualize formant data, providing publication-ready results across platforms that support interactive, reproducible workflows.
 
-VowSpace has a modular architecture consisting of a main interface and two specialized tools: Audio Analysis and DataFrame Editor. The Audio Analysis window serves as a primary data extraction point, allowing users to load audio files in supported formats (i.e., .aiff, .aif, .aifc, .flac, .wav, .mp3, .ogg), visualize them as spectrograms, and extract acoustic features such as intensity, pitch, and vowel formants (F1–F4) using the Parselmouth library [@parselmouth2018]. Users can also add extracted formant values directly to the visualizer by clicking on the generated spectrogram, which minimizes manual transcription errors and facilitates immediate validation against existing distributions, thereby streamlining the transition from raw signal to analyzed data. These spectrogram images can be saved as JPEG, PNG, or TIFF files with visualized acoustic/prosodic information.
+VowSpace has a modular architecture consisting of a main interface and two specialized tools: Audio Analysis and DataFrame Editor. The Audio Analysis window serves as a primary data extraction point, allowing users to load audio files in supported formats (i.e., .aiff, .aif, .aifc, .flac, .wav, .mp3, .ogg), visualize them as spectrograms, and extract acoustic features such as intensity, pitch, and vowel formants (F1–F5) using the Parselmouth library [@parselmouth2018]. Users can also add extracted formant values directly to the visualizer by clicking on the generated spectrogram, which minimizes manual transcription errors and facilitates immediate validation against existing distributions, thereby streamlining the transition from raw signal to analyzed data. These spectrogram images can be saved as JPEG, PNG, or TIFF files with visualized acoustic/prosodic information.
 
 The built-in DataFrame Editor allows in-app data editing with instant scatterplot updates upon saving to ensure that visualization remains synchronized with the underlying dataset. Data can be exported as a spreadsheet. VowSpace uses a minimal input format to stay compatible with other tools and to fit easily into phonetic research workflows without requiring complex data structures. The supported minimal format is as follows:
 
