### 2026-04-22 #10189 @alicagank
@RichardLitt, I should be able to cover all that tonight (which should be daytime in New Zealand) I presume. :)

Thank you very much!
