### 2026-04-22 #10189 @alicagank
Hi again @RichardLitt, it looks like I am still having some issues with the macOS build, I am overseas for a conference, and testing the builds properly across all systems takes a bit of time. Would it be alright if I do a little more testing before creating the release? It doesn't seem to be anything regarding the app, but I might have to reconfigure the Nuitka build workflow. I'll be working on that asap. Thank you very much for understanding.
