### 2026-04-23 47b02f4
Consistency, capitalization of "F" in the UI.
