### 2026-04-23 7c0e958
Some error handling
