### 2026-04-23 5d3b565
macOS fix PyQt5
