### 2026-04-23 52bf7e2
macOS fix PyQt5
