### 2026-04-23 8b72ed3
add version metadata
