### 2026-04-23 5d9344d
Rewrite accoring to Nuitka guide
