### 2026-04-23 9dfdd80
Added QIcon
