### 2026-04-23 #10189 @alicagank
Hi, @RichardLitt I have tried to find a workaround with the Nuitka workflow, but apparently Nuitka’s PyQt5 support is [very limited](https://nuitka.net/pages/pyqt5.html), which seems to be why the Nuitka build is not working reliably on macOS. VowSpace currently uses PyQt5 for its main UI, I am considering migrating to PyQt6 in a future release, probably within the next two months.

Since this workflow is not a core component of the application, would it be acceptable to compile the macOS executable with PyInstaller instead of Nuitka for this release, and remove the macOS Nuitka build from the workflow for now? I’m very open to suggestions.
