### 2026-04-24 #10189 @RichardLitt
Right.

Well, it's not necessary for publication. You don't have to have a feature complete OSS package to publish in JOSS, and I think it would be a failing of OSS if you did.

So, I would encourage just closing up this submission as is.  :)
