### 2026-04-24 8988945
Bump version to 1.4.3 for JOSS release
