+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "VowSpace"
-version = "1.4.2"
+version = "1.4.3"
 description = "Vowel Space Visualizer"
 authors = [{ name = "VowSpace Authors" }]
 readme = "README.md"
