### 2026-04-24 777c7ac
Bump version to 1.4.4 for JOSS release
