### 2026-04-24 #10189 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
