### 2026-04-27 #10189 @samhforbes
I can see you've already discussed the missing DOI, so looks like this is all good to go
