### 2026-04-27 #10189 @samhforbes
> Sorry for the delay, @samhforbes, I was on a flight. Everything looks fine on my end, and thank you very much for your help.
>
> Editor Richard Littauer and reviewer hadware made some pull requests during the review, which must be the source of the additional authors on Zenodo. I wasn’t sure whether it would be appropriate to remove their names. Should I (or can I) remove them from the author list on Zenodo?

I would normally suggest removing these so it matches, honestly
