### 2026-04-27 #10189 @alicagank
Thank you for the clarification @samhforbes, that makes sense. I’d prefer to quickly check with the editor @RichardLitt and the reviewer @hadware before removing their names from the Zenodo author list, just to make sure they’re okay with it. I hope that’s alright. I’ll take care of that once I’ve received confirmation. :)
