+++ CITATION.cff
@@ -1,0 +1,28 @@
+cff-version: "1.2.0"
+authors:
+- family-names: Kaya
+  given-names: Ali Çağan
+  orcid: "https://orcid.org/0009-0007-9386-9766"
+doi: 10.5281/zenodo.19718178
+message: If you use this software, please cite our article in the
+  Journal of Open Source Software.
+preferred-citation:
+  authors:
+  - family-names: Kaya
+    given-names: Ali Çağan
+    orcid: "https://orcid.org/0009-0007-9386-9766"
+  date-published: 2026-04-29
+  doi: 10.21105/joss.10189
+  issn: 2475-9066
+  issue: 120
+  journal: Journal of Open Source Software
+  publisher:
+    name: Open Journals
+  start: 10189
+  title: "VowSpace: A vowel formant analysis application for phonetic
+    research"
+  type: article
+  url: "https://joss.theoj.org/papers/10.21105/joss.10189"
+  volume: 11
+title: "VowSpace: A vowel formant analysis application for phonetic
+  research"
